% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_map.R
\name{alpha_map}
\alias{alpha_map}
\title{An Alpha shape mapping function}
\usage{
alpha_map(lon_lat, alpha = 0.5)
}
\arguments{
\item{lon_lat}{A two column dataframe or matrix with spatial coordinates}

\item{alpha}{The alpha parameter for the package alphahull}
}
\description{
This function takes as series of lon lat points and returns a data frame with the coordinates and group of an alpha shape
}
\keyword{alpha}
\keyword{map}
\keyword{shape}
\keyword{spatial}
