% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atan_sort.R
\name{atan_sort}
\alias{atan_sort}
\title{A helper function for ordering segments by atan
This function returns ordered segments for a polygon}
\usage{
atan_sort(df)
}
\arguments{
\item{df}{A two column dataframe or matrix with spatial coordinates}
}
\description{
A helper function for ordering segments by atan
This function returns ordered segments for a polygon
}
